+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/validators/ArchiveValidationHandler.java
@@ -1,10 +1,9 @@
-/* Copyright 2021 Radix Publishing Ltd incorporated in Jersey (Channel Islands).
- *
+/*
+ * Copyright 2021 Radix DLT Ltd incorporated in England.
  * Licensed under the Radix License, Version 1.0 (the "License"); you may not use this
  * file except in compliance with the License. You may obtain a copy of the License at:
  *
  * radixfoundation.org/licenses/LICENSE-v1
- *
  * The Licensor hereby grants permission for the Canonical version of the Work to be
  * published, distributed and used under or by reference to the Licensor’s trademark
  * Radix ® and use of any unregistered trade names, logos or get-up.
@@ -62,15 +61,13 @@
  * permissions under this License.
  */
 
-package com.radixdlt.api.handler;
+package com.radixdlt.api.validators;
 
 import com.radixdlt.networks.Addressing;
 import org.json.JSONObject;
 
 import com.google.inject.Inject;
-import com.google.inject.Singleton;
 import com.radixdlt.api.data.ValidatorInfoDetails;
-import com.radixdlt.api.service.ValidatorArchiveInfoService;
 import com.radixdlt.crypto.ECPublicKey;
 import com.radixdlt.utils.functional.Result;
 
@@ -87,7 +84,6 @@ import static com.radixdlt.api.data.ApiErrors.INVALID_PAGE_SIZE;
 import static com.radixdlt.utils.functional.Result.allOf;
 import static com.radixdlt.utils.functional.Result.ok;
 
-@Singleton
 public class ArchiveValidationHandler {
 	private final ValidatorArchiveInfoService validatorInfoService;
 	private final Addressing addressing;
