+++ radixdlt-core/radixdlt/src/test/java/com/radixdlt/api/handler/ArchiveValidationHandlerTest.java
@@ -64,13 +64,14 @@
 package com.radixdlt.api.handler;
 
 import com.radixdlt.api.store.ValidatorUptime;
+import com.radixdlt.api.validators.ArchiveValidationHandler;
 import com.radixdlt.networks.Addressing;
 import com.radixdlt.networks.Network;
 import org.json.JSONObject;
 import org.junit.Test;
 
 import com.radixdlt.api.data.ValidatorInfoDetails;
-import com.radixdlt.api.service.ValidatorArchiveInfoService;
+import com.radixdlt.api.validators.ValidatorArchiveInfoService;
 import com.radixdlt.crypto.ECKeyPair;
 import com.radixdlt.crypto.ECPublicKey;
 import com.radixdlt.identifiers.REAddr;
